// api-tests/generated/models/IssueTypeSchemeProjectAssociation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* eslint-disable */
/**
 * Details of the association between an issue type scheme and project.
 */
export type IssueTypeSchemeProjectAssociation = {
  /**
   * The ID of the issue type scheme.
   */
  issueTypeSchemeId: string;
  /**
   * The ID of the project.
   */
  projectId: string;
};